+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2021-12-14
 ### Added
 - add confirmation prompt to installer
+- add install process to installer
 
 ## 2021-12-13
 ### Added